+++ web/blueprint/src/lib/components/concepts/ConceptView.svelte
@@ -1,8 +1,12 @@
 <script lang="ts">
+  import {goto} from '$app/navigation';
   import {editConceptMutation} from '$lib/queries/conceptQueries';
   import {queryAuthInfo} from '$lib/queries/serverQueries';
   import {queryEmbeddings} from '$lib/queries/signalQueries';
-  import type {Concept} from '$lilac';
+  import {createDatasetViewStore} from '$lib/stores/datasetViewStore';
+  import {datasetLink} from '$lib/utils';
+  import {serializePath, type Concept} from '$lilac';
+  import {Button, ToastNotification} from 'carbon-components-svelte';
   import {ViewOff} from 'carbon-icons-svelte';
   import ThumbsDownFilled from 'carbon-icons-svelte/lib/ThumbsDownFilled.svelte';
   import ThumbsUpFilled from 'carbon-icons-svelte/lib/ThumbsUpFilled.svelte';
@@ -11,6 +15,7 @@
   import ConceptExampleList from './ConceptExampleList.svelte';
   import ConceptMetrics from './ConceptMetrics.svelte';
   import ConceptPreview from './ConceptPreview.svelte';
+  import DatasetFieldEmbeddingSelector from './DatasetFieldEmbeddingSelector.svelte';
   import ConceptLabeler from './labeler/ConceptLabeler.svelte';
 
   export let concept: Concept;
@@ -25,6 +30,36 @@
   $: negativeExamples = Object.values(concept.data).filter(v => v.label == false);
 
   $: randomPositive = positiveExamples[Math.floor(Math.random() * positiveExamples.length)];
+
+  // Apply to a dataset.
+  let applyDataset: {namespace: string; name: string} | undefined | null = undefined;
+  let applyPath: string[] | undefined;
+  let applyEmbedding: string | undefined = undefined;
+  $: datasetViewStore =
+    applyDataset != null && applyPath != null && applyEmbedding != null
+      ? createDatasetViewStore(applyDataset.namespace, applyDataset.name)
+      : null;
+  function openDataset() {
+    const pathId = applyPath ? serializePath(applyPath) : undefined;
+    if (
+      pathId == null ||
+      applyEmbedding == null ||
+      applyDataset == null ||
+      datasetViewStore == null ||
+      $datasetViewStore == null
+    ) {
+      return;
+    }
+
+    datasetViewStore.addSearch({
+      path: [pathId],
+      type: 'concept',
+      concept_namespace: concept.namespace,
+      concept_name: concept.concept_name,
+      embedding: applyEmbedding
+    });
+    goto(datasetLink(applyDataset.namespace!, applyDataset.name!, $datasetViewStore));
+  }
 
   function remove(id: string) {
     if (!concept.namespace || !concept.concept_name) return;
@@ -59,6 +94,36 @@
   <Expandable expanded>
     <div slot="above" class="text-md font-semibold">Try it</div>
     <ConceptPreview example={randomPositive} {concept} slot="below" />
+  </Expandable>
+
+  <Expandable>
+    <div slot="above" class="text-md font-semibold">Apply to a dataset</div>
+    <div slot="below">
+      <DatasetFieldEmbeddingSelector
+        bind:dataset={applyDataset}
+        bind:path={applyPath}
+        bind:embedding={applyEmbedding}
+      />
+      {#if applyDataset != null && applyPath != null && applyEmbedding != null}
+        <div class="mt-4">
+          <Button
+            class=""
+            iconDescription={'Open dataset and apply concept.'}
+            on:click={() => openDataset()}>Search by concept</Button
+          >
+        </div>
+      {:else}
+        <ToastNotification
+          hideCloseButton
+          kind="warning"
+          fullWidth
+          lowContrast
+          title="Choose a dataset with a computed embedding"
+          caption={'Dataset has no fields with computed embeddings. ' +
+            'Please compute an embedding index before you can search by concept.'}
+        />
+      {/if}
+    </div>
   </Expandable>
 
   {#if $embeddings.data}
